Add npm run-script build
* split webspack into two configs
* dev config is still default
* use webpack from local install
